symbols working for codeblock toolbars
